Create a repo for tests on flag mixing
