+++ var/spack/repos/flags.test/packages/y/package.py
@@ -1,0 +1,12 @@
+# Copyright 2013-2024 Lawrence Livermore National Security, LLC and other
+# Spack Project Developers. See the top-level COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: (Apache-2.0 OR MIT)
+from spack.package import *
+
+
+class Y(Package):
+    version("2.1")
+    version("2.0")
+
+    depends_on("c", type="build")
