+++ var/spack/repos/flags.test/repo.yaml
@@ -1,0 +1,2 @@
+repo:
+  namespace: flags.test
